LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/animated/observe-animated-image-webp.tentative-expected.txt



FAIL Same origin animated image is observable and has a first frame. assert_less_than: renderTime should be smaller than the delay applied to the second frame expected a number less than 2000 but got 3076
